Improve ipld-cbor
