+++ core/ipld/cbor/Cargo.toml
@@ -25,9 +25,14 @@ block-format = { package = "rust-block-format", path = "../../block-format" }
 cid = { package = "rust-cid", path = "../../cid" }
 
 [dev-dependencies]
+criterion = "0.3"
 util = { package = "rust-ipfs-util", path = "../../util" }
 hex-literal = "0.2.1"
 
 [features]
-bench = []
 bigint = [ "num-bigint" ]
+
+[[bench]]
+name = "benchmarks"
+path = "benches/benchmarks.rs"
+harness = false
