+++ core/ipld/cbor/src/bigint.rs
@@ -20,24 +20,24 @@
 // Thus, we just provide `num::BigUint` to stand for `big.NewInt` in rust. if someone need real
 // `num::BigInt`, it should implement by himself.
 
-use std::fmt::{Display, Error, Formatter};
+use std::fmt;
 
 use num_bigint::BigUint;
-use serde::{Deserialize, Deserializer, Serialize, Serializer};
 
-#[derive(Clone, Debug, Hash, Eq, PartialEq, Ord, PartialOrd)]
+/// A CBOR implementation of big int.
+#[derive(PartialEq, Eq, PartialOrd, Ord, Clone, Debug, Hash)]
 pub struct CborBigUint(pub BigUint);
 
-impl Display for CborBigUint {
-    fn fmt(&self, f: &mut Formatter<'_>) -> Result<(), Error> {
+impl fmt::Display for CborBigUint {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         self.0.fmt(f)
     }
 }
 
-impl Serialize for CborBigUint {
+impl serde::Serialize for CborBigUint {
     fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
     where
-        S: Serializer,
+        S: serde::Serializer,
     {
         // Returns the byte representation of the `BigUint` in big-endian byte order
         let v = self.0.to_bytes_be();
@@ -46,12 +46,86 @@ impl Serialize for CborBigUint {
     }
 }
 
-impl<'de> Deserialize<'de> for CborBigUint {
+impl<'de> serde::Deserialize<'de> for CborBigUint {
     fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
     where
-        D: Deserializer<'de>,
+        D: serde::Deserializer<'de>,
     {
         let v = serde_bytes::ByteBuf::deserialize(deserializer)?;
         Ok(CborBigUint(BigUint::from_bytes_be(&v)))
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::collections::BTreeMap;
+
+    use block_format::Block;
+    use multihash::Hash;
+    use serde::{Deserialize, Serialize};
+
+    use super::CborBigUint;
+    use crate::localcid::CborCid;
+    use crate::node::IpldNode;
+    use crate::obj::{Obj, SortedStr};
+
+    #[test]
+    fn test_cid_and_bigint() {
+        #[derive(PartialEq, Eq, Debug, Serialize, Deserialize)]
+        struct Foo {
+            big_int: CborBigUint,
+            cid: CborCid,
+        }
+
+        let node = IpldNode::from_obj(Obj::Null, Hash::SHA2256).unwrap();
+        let cid = node.cid().clone();
+        let foo1 = Foo {
+            big_int: CborBigUint(1_u64.into()),
+            cid: CborCid(cid),
+        };
+
+        let bytes = serde_cbor::to_vec(&foo1).unwrap();
+        let foo2 = serde_cbor::from_slice::<Foo>(&bytes).unwrap();
+        assert_eq!(foo1, foo2);
+    }
+
+    #[test]
+    fn test_bigint_roundtrip() {
+        #[derive(PartialEq, Eq, Debug, Serialize, Deserialize)]
+        struct TestMe {
+            hello: CborBigUint,
+            world: CborBigUint,
+            hi: i32,
+        }
+        let me1 = TestMe {
+            hello: CborBigUint(100_u64.into()),
+            world: CborBigUint(99_u64.into()),
+            hi: 0,
+        };
+        let bytes = serde_cbor::to_vec(&me1).unwrap();
+        let me2 = serde_cbor::from_slice::<TestMe>(&bytes).unwrap();
+        assert_eq!(me1, me2);
+
+        type M = BTreeMap<SortedStr, TestMe>;
+        let mut m1 = M::new();
+        m1.insert(
+            "hello".into(),
+            TestMe {
+                hello: CborBigUint(10_u64.into()),
+                world: CborBigUint(101_u64.into()),
+                hi: 1,
+            },
+        );
+        m1.insert(
+            "world".into(),
+            TestMe {
+                hello: CborBigUint(9_u64.into()),
+                world: CborBigUint(901_u64.into()),
+                hi: 3,
+            },
+        );
+        let bytes = serde_cbor::to_vec(&m1).unwrap();
+        let m2: M = serde_cbor::from_slice(&bytes).unwrap();
+        assert_eq!(m1, m2);
     }
 }
