+++ core/ipld/cbor/src/localcid.rs
@@ -1,48 +1,40 @@
 // Copyright 2019-2020 PolkaX. Licensed under MIT or Apache-2.0.
 
-use std::borrow::Borrow;
-use std::ops::{Deref, DerefMut};
-
-use serde::de::{Deserialize, Deserializer, Error};
-use serde::ser::{Serialize, Serializer};
+use cid::Cid;
+use serde::{
+    de::{Deserialize, Deserializer, Error},
+    ser::{Serialize, Serializer},
+};
 use serde_cbor::tags::Tagged;
 
 use crate::error::IpldCborError;
-use crate::Cid;
-
-#[derive(PartialEq, Eq, Clone, Debug)]
-pub struct CborCid(pub Cid);
 
 /// CID_CBOR_TAG is the integer used to represent cid tags in CBOR.
-pub const CID_CBOR_TAG: u64 = 42;
+pub(crate) const CID_CBOR_TAG: u64 = 42;
+
+/// A CID Wrapper that implements CBOR serialization/deserialization.
+#[derive(PartialEq, Eq, Clone, Debug)]
+pub struct CborCid(pub(crate) Cid);
+
+impl CborCid {
+    /// Convert to inner cid.
+    pub fn into_inner(self) -> Cid {
+        self.0
+    }
+}
 
 impl Serialize for CborCid {
-    fn serialize<S>(&self, s: S) -> Result<S::Ok, S::Error>
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
     where
         S: Serializer,
     {
         // add 0 at start
-        let mut b = vec![0_u8];
-        b.extend(self.0.to_bytes());
+        let mut bytes = vec![0_u8];
+        bytes.extend(self.0.to_bytes());
         // Special process for bytes, due to serde_cbor would treat Vec<u8> as Array u8, not bytes
-        let value = serde_bytes::Bytes::new(&b);
-        Tagged::new(Some(CID_CBOR_TAG), value).serialize(s)
+        let value = serde_bytes::Bytes::new(&bytes);
+        Tagged::new(Some(CID_CBOR_TAG), value).serialize(serializer)
     }
-}
-
-pub fn deserialize_cid_from_bytes(res: &[u8]) -> Result<Cid, IpldCborError> {
-    if res.len() == 0 {
-        return Err(IpldCborError::DeserializeCid(format!("Value was empty")));
-    }
-
-    if res[0] != 0 {
-        return Err(IpldCborError::DeserializeCid(format!(
-            "Invalid multibase on IPLD link"
-        )));
-    }
-
-    let cid = Cid::from(&res[1..])?;
-    Ok(cid)
 }
 
 impl<'de> Deserialize<'de> for CborCid {
@@ -53,9 +45,7 @@ impl<'de> Deserialize<'de> for CborCid {
         let tagged = Tagged::<serde_bytes::ByteBuf>::deserialize(deserializer)?;
         match tagged.tag {
             Some(CID_CBOR_TAG) | None => {
-                let res = tagged.value.to_vec();
-
-                let cid = deserialize_cid_from_bytes(&res)
+                let cid = deserialize_cid_from_bytes(&tagged.value.to_vec())
                     .map_err(|e| D::Error::custom(format!("Cid deserialize failed: {:}", e)))?;
                 Ok(CborCid(cid))
             }
@@ -64,15 +54,36 @@ impl<'de> Deserialize<'de> for CborCid {
     }
 }
 
+pub(crate) fn deserialize_cid_from_bytes(bytes: &[u8]) -> Result<Cid, IpldCborError> {
+    if bytes.is_empty() {
+        return Err(IpldCborError::DeserializeCid("Value was empty".to_string()));
+    }
+
+    if bytes[0] != 0 {
+        return Err(IpldCborError::DeserializeCid(
+            "Invalid multibase on IPLD link".to_string(),
+        ));
+    }
+
+    let cid = Cid::from(&bytes[1..])?;
+    Ok(cid)
+}
+
 impl From<Cid> for CborCid {
     fn from(cid: Cid) -> Self {
         CborCid(cid)
     }
 }
 
-impl Borrow<Cid> for CborCid {
+impl std::borrow::Borrow<Cid> for CborCid {
     fn borrow(&self) -> &Cid {
         &self.0
+    }
+}
+
+impl std::borrow::BorrowMut<Cid> for CborCid {
+    fn borrow_mut(&mut self) -> &mut Cid {
+        &mut self.0
     }
 }
 
@@ -82,7 +93,13 @@ impl AsRef<Cid> for CborCid {
     }
 }
 
-impl Deref for CborCid {
+impl AsMut<Cid> for CborCid {
+    fn as_mut(&mut self) -> &mut Cid {
+        &mut self.0
+    }
+}
+
+impl std::ops::Deref for CborCid {
     type Target = Cid;
 
     fn deref(&self) -> &Self::Target {
@@ -90,7 +107,7 @@ impl Deref for CborCid {
     }
 }
 
-impl DerefMut for CborCid {
+impl std::ops::DerefMut for CborCid {
     fn deref_mut(&mut self) -> &mut Self::Target {
         &mut self.0
     }
