+++ core/ipld/cbor/tests/test_objects/empty-array.json
