+++ core/ipld/cbor/tests/test_objects/empty-obj.cbor
