+++ core/ipld/cbor/tests/test_objects/foo.cbor
