+++ core/ipld/cbor/tests/test_objects/foo2.cbor
