+++ core/ipld/cbor/tests/test_objects/obj-no-link.json
