+++ core/ipld/cbor/tests/test_objects/obj-with-link.cbor
